// src/emotion-core/types/NRCEmotion.ts
/**
 * NRC Emotion Dimensions
 * Source: NRC Emotion Lexicon v0.92
 *
 * NOTE:
 * - POSITIVE / NEGATIVE are sentiment dimensions
 * - Others are discrete emotions
 */
export enum NRCEmotion {
  ANGER = 'anger',
  ANTICIPATION = 'anticipation',
  DISGUST = 'disgust',
  FEAR = 'fear',
  JOY = 'joy',
  SADNESS = 'sadness',
  SURPRISE = 'surprise',
  TRUST = 'trust',
  POSITIVE = 'positive',
  NEGATIVE = 'negative',
}